moved autoFit functionality to new component #606
